vm: Add bind() and jump() to CodeGenerator.
Untested code. The plan is to modify the byteode to used structured
control flow opcodes to simplify the writing of a one-pass (baseline)
compiler.

+++ 2016/vm/vm.h
@@ -35,7 +35,7 @@
 //
 
 using Bytecode = std::vector<int32_t>;
-using MachineCode = std::vector<uint32_t>;
+using MachineCode = std::vector<uint8_t>;
 
 // Assemble the assembly code in the stream provided by |in| into bytecode.
 //
